Create brain-teaser 1
A software developer who has built over ten(10) mostly viewed website was on the internet decides to push his files in a reputable hosting platforms. The developer makes use of three different CSS files. The first contains the font-family, while the second contains the background, and the referencing contains body and paragraph. The developer uploads all the necessary files on the server and renders high quality web based files. inside the developer server consist of the html-5 files with a <doctype> and uses ten(10) different html tags by default. As a developer kindly render this file tags correctly.